correction in size of the pictures(htmconfig.json "input Dimensions")
ommition of the part of encoder which considers the colors of the pictures too (because our pics are black and white)
putting triangle instead of the angleCross in input folder.

back then when the angleCross has also been a part of the entry, sometimes the similarity of the angleCross and straightCross, was more tha the similarity of these categories to themselves. but it was completely normal, because angleCrosses are also straightCrosses which are rotated by 45 degrees.
so we can say that the ImageClassifier was actually doing a good work, and was also able to find similarities, between the rotated shapes too